pyproject.toml
[project]
name = "cbot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "dotenv>=0.9.9",
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
    "requests>=2.32.5",
    "click>=8.1.7",
    "demeter-fetch>=1.3.7",
    "google-cloud-bigquery>=3.36.0",
    "google-cloud-bigquery-storage>=2.27.0",
    "web3>=7.13.0",
    "eth-abi>=5.2.0",
    "eth-hash>=0.7.1",
    "polars>=1.33.1",
    "torch>=2.8.0",
    "torch-geometric>=2.6.1",
    "torch-scatter>=2.1.2",
    "torch-sparse>=0.6.18",
    "torch-cluster>=1.6.3",
    "torch-spline-conv>=1.2.2",
    "scikit-learn>=1.7.2",
    "torch-geometric-temporal[index]>=0.56.2",
    "dask>=2025.9.1",
    "node2vec>=0.5.0",
    "networkx>=3.5",
    "matplotlib>=3.10.6",
    "umap-learn>=0.5.9.post2",
]

[dependency-groups]
dev = [
    "ipython>=9.5.0",
    "jupyter>=1.1.1",
    "mypy>=1.17.1",
    "networkx-stubs>=0.0.1",
    "pandas-stubs>=2.3.2.250827",
    "ruff>=0.12.10",
    "types-google-cloud-ndb>=2.3.0.20250317",
    "types-requests>=2.32.4.20250809",
]

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]


[tool.uv]
find-links = [
    "https://data.pyg.org/whl/torch-2.8.0+cpu.html",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true


[[tool.mypy.overrides]]
module = ["torch_geometric_temporal.*"]
ignore_missing_imports = true

[tool.ruff]
# Set target Python version (match your project’s Python version, e.g., 3.11)
target-version = "py311"
# Set line length (same as Black’s default)
line-length = 88
indent-width = 4
# Include commonly ignored directories
# select = ["E", "F", "W", "I", "N", "D"] # Example: pycodestyle, pyflakes, warnings, imports, naming, docstrings

exclude = [
    ".venv",
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
# Enable all rules for strict linting
select = ["ALL"]
# Optionally ignore specific rules if they’re too strict for your project
ignore = [
    "S101", # Allow using assert
    "T201", # Allow print
    # "PLR0913", # Allow > 5 arguments to a function
    #     "D203", # Example: Ignore specific docstring rules if needed
    #     "D213", # Example: Ignore conflicting docstring rules
]
# Allow automatic fixing for all fixable rules
fixable = ["ALL"]
# Allow unused variables with underscore prefix
# dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
# Enforce strict line length
# line-too-long = "E501"

[tool.ruff.format]
# Use Black-compatible formatting
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.mypy]
python_version = "3.11" # Or your target Python version
strict = true
# warn_unused_ignores = true
# disallow_untyped_defs = true
# disallow_incomplete_defs = true
# check_untyped_defs = true
# no_implicit_optional = true
# warn_redundant_casts = true
# warn_unused_configs = true
